+++ backend/controllers/request.go
@@ -55,13 +55,19 @@ func RequestToServiceAnalytics(ctx *gin.Context, service models.Service, inputDa
 
 func RequestToServiceInnovation(ctx *gin.Context, service models.Service, inputData models.ServiceRequestInput) {
 	var err error
+	var serviceData models.ServiceRequestResultData
 
 	requestData := inputData.Data
 	additionalParams, _ := json.Marshal(requestData.AdditionalParams)
 	postBody := []byte(fmt.Sprintf(`{ "additional_params": %v , "images":  [ "%v" ]}`,
 		string(additionalParams), strings.Join(requestData.Images, `", "`)))
 
-	serviceData, err := RequestToInnovationSync(postBody, service.Slug)
+	if service.Slug == "face-occlusion-attribute" {
+		RequestToFaceOcclusionAttribute(ctx, postBody)
+		return
+	} else {
+		serviceData, err = RequestToInnovationSync(postBody, service.Slug)
+	}
 
 	if err != nil {
 		log.WithFields(log.Fields{